requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2
pydantic-settings
flake8
autopep8